Move error functions to nn.loss
